added sweetalert actions

+++ src/actions/crudActions.js
@@ -1,3 +1,5 @@
+import Swal from 'sweetalert2'
+
 import { 
     createOneUserFetch, 
     deleteOneUserFetch, 
@@ -18,9 +20,10 @@ const getAllUsers = (users) => {
 export const getAllUsersAction = () => {
     return async (dispatch) => {
         const resp = await getAllUsersFetch();
-        dispatch(getAllUsers(resp.users));
+        dispatch(getAllUsers(resp));
     }
 };
+
 // GET ONE USER
 const getOneUser = (user) => {
     return {
@@ -34,6 +37,7 @@ export const getOneUserAction = (id) => {
         dispatch(getOneUser(resp))
     }
 }
+
 // CREATE ONE USER
 const createOneUser = (user) => {
     return {
@@ -44,9 +48,18 @@ const createOneUser = (user) => {
 export const createOneUserAction = (user) => {
     return async (dispatch) => {
         const resp = await createOneUserFetch(user);
+        Swal.fire({
+            title: "User",
+            icon: "success",
+            text: `The User has been created`,
+            showConfirmButton: false,
+            position: "bottom-left",
+            timer: 1500
+        })
         dispatch(createOneUser(resp));
     };
 };
+
 // DELETE ONE USER
 const deleteOneUser = (id) => {
     return {
@@ -56,10 +69,24 @@ const deleteOneUser = (id) => {
 };
 export const deleteOneUserAction = (id) => {
     return async (dispatch) => {
-        const resp = await deleteOneUserFetch(id)
-        dispatch(deleteOneUser(resp._id))
+        
+        const swal = await Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          });
+          if(swal.isConfirmed){
+            const resp = await deleteOneUserFetch(id);
+            dispatch(deleteOneUser(resp._id))
+          };
     };
 };
+
+// Edit ONE USER
 const editOneUser = (user) => {
     return {
         type: TYPES.editOneUser,
@@ -68,7 +95,15 @@ const editOneUser = (user) => {
 };
 export const editOneUserAction = (user) => {
     return async (dispatch) => {
-        const resp = editOneUserFetch(user);
+        const resp = await editOneUserFetch(user);
+        Swal.fire({
+            icon: "success",
+            title: "User",
+            text: "The User has been edited",
+            showConfirmButton: false,
+            position: "bottom-left",
+            timer: 1500
+        })
         dispatch(editOneUser(resp))
     };
 };
